updated dem hexify
