+++ community/sina/DEM_Tile_Hexify/meta.json
@@ -44,11 +44,23 @@
             "fused:udfType": "auto",
             "fused:slug": "DEM_Tile_Hexify",
             "fused:name": "DEM_Tile_Hexify",
-            "fused:id": null,
-            "fused:gitUrl": "https://github.com/fusedio/udfs/tree/743a6f777bca66209cc8c66b868f8a9ca9c59efe/public/DEM_Tile_Hexify/",
-            "fused:gitShortUrl": "https://github.com/fusedio/udfs/tree/743a6f7/public/DEM_Tile_Hexify/",
-            "fused:gitPath": "public/DEM_Tile_Hexify",
-            "fused:gitRef": "743a6f777bca66209cc8c66b868f8a9ca9c59efe",
+            "fused:gitPath": "community/sina/DEM_Tile_Hexify",
+            "fused:gitRef": "2528576155696ad0aa24beef28643891343f2211",
+            "fused:gitAuthorNames": [
+              null
+            ],
+            "fused:gitAuthorUsernames": [
+              "MaxLenormand"
+            ],
+            "fused:gitAuthorUrls": [
+              "https://github.com/MaxLenormand"
+            ],
+            "fused:gitAuthorAvatarUrls": [
+              "https://avatars.githubusercontent.com/u/44133329?v=4"
+            ],
+            "fused:gitLastModified": "2025-07-23T14:53:34+00:00",
+            "fused:explorerTab": "self",
+            "fused:gitRepo": "fusedio/udfs",
             "fused:assetUrl": "https://fused-magic.s3.us-west-2.amazonaws.com/thumbnails/udfs-staging/DEM_Tile_Hexify.png",
             "fused:tags": [
               {
@@ -66,33 +78,11 @@
             ],
             "fused:description": "#### Overview\nThis example demonstrates how terrain TIFF tiles can be dynamically converted into an H3 grid. The raster values are aggregated (sum or mean) for H3 hexes at a resolution dependent on the zoom level. The terrain data come from diverse global sources, including ArcticDEM funded by the NSF, Geoscience Australia, Austria’s Digital Terrain Model, Canada's Open Government License, EU Copernicus data, NOAA's ETOPO1, INEGI in Mexico, New Zealand's Land Information authority, Kartverket in Norway, the UK's Environment Agency, and the U.S. Geological Survey for 3DEP and SRTM. \n\n## External links\n\n- [AWS open data](https://registry.opendata.aws/terrain-tiles/)\n- [Attribution link](https://github.com/tilezen/joerd/blob/master/docs/attribution.md)\n",
             "fused:explorerPin": "9",
-            "fused:gitAuthorNames": [
-              "Sina Kashuk",
-              "Isaac Brodsky"
-            ],
-            "fused:gitAuthorUsernames": [
-              "kashuk",
-              "isaacbrodsky"
-            ],
-            "fused:gitAuthorUrls": [
-              "https://github.com/kashuk",
-              "https://github.com/isaacbrodsky"
-            ],
-            "fused:gitAuthorAvatarUrls": [
-              "https://avatars.githubusercontent.com/u/12432685?v=4",
-              "https://avatars.githubusercontent.com/u/9139378?v=4"
-            ],
-            "fused:gitLastModified": "2024-12-20T00:09:37+00:00",
-            "fused:explorerTab": "public",
-            "fused:gitRepo": "fusedio/udfs"
+            "fused:userCreatedShareToken": "fsh_6Vrv8Jqg7KrUZFmEPolOME",
+            "fused:id": null
           },
           "source": "DEM_Tile_Hexify.py",
-          "headers": [
-            {
-              "module_name": "utils",
-              "source_file": "utils.py"
-            }
-          ]
+          "headers": []
         }
       }
     ],
